extracted validators from controller functions
