+++ server/routes/user.js
@@ -1,8 +1,12 @@
 const express = require("express");
 const router = express.Router();
+const {
+  userValidationRules,
+  validate,
+} = require("../validators/userValidators");
 const userController = require("../controllers/userController");
 
-router.post("/create", userController.create);
+router.post("/create", userValidationRules(), validate, userController.create);
 router.post("/login", userController.login);
 
 module.exports = router;
